Foundation: fix some #include "Foundation/header.h" because Xcode 15.3 complains that it is not <Foundation/header.h>

+++ Foundation/Sources/NSXMLParser.m
@@ -316,7 +316,7 @@ static NSDictionary *entitiesTable;
 		}
 	else if([d length] > 0)
 		{ // append to new buffer
-			unsigned cpoff=cp-(char *)[buffer bytes];	// get current offset
+			long cpoff=cp-(char *)[buffer bytes];	// get current offset
 			if(![buffer isKindOfClass:[NSMutableData class]])
 				{ // make a mutable copy
 					NSData *b=buffer;				// remember
